+++ samples/net/lwm2m_client/src/modules.h
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2022 Nordic Semiconductor
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef _MODULES_H
+#define _MODULES_H
+
+int init_led_device(void);
+void init_timer_object(void);
+void init_temp_sensor(void);
+void init_firmware_update(void);
+
+#endif
